+++ src/providers/BiconomyProvider.tsx
@@ -3,41 +3,44 @@ import {
   type Hex,
   type SupportedSigner,
   createSmartAccountClient,
-  type BiconomySmartAccountV2Config
-} from "@biconomy/account"
-import type { QueryClient } from "@tanstack/react-query"
+  type BiconomySmartAccountV2Config,
+} from "@biconomy/account";
 import {
-  type ReactNode,
-  createContext,
-  useEffect,
-  useMemo,
-  useState
-} from "react"
-import { useWalletClient } from "wagmi"
+  useQuery,
+  UseQueryResult,
+  type QueryClient,
+} from "@tanstack/react-query";
+import { type ReactNode, createContext, useMemo } from "react";
+import { useWalletClient } from "wagmi";
 
-export type {BiconomySmartAccountV2} from "@biconomy/account"
+export type { BiconomySmartAccountV2 } from "@biconomy/account";
 
 export type BiconomyProviderProps = {
   /** The children of the provider */
-  children: ReactNode
+  children: ReactNode;
   /** The Biconomy configuration */
-  config: Partial<BiconomySmartAccountV2Config> & { bundlerUrl: string, biconomyPaymasterApiKey: string}
+  config: Partial<BiconomySmartAccountV2Config> & {
+    bundlerUrl: string;
+    biconomyPaymasterApiKey: string;
+  };
   /** The Tanstack Query client instance */
-  queryClient: QueryClient | undefined
-}
+  queryClient: QueryClient | undefined;
+};
 
 export type BiconomyContextPayload = {
   /** The BiconomySmartAccountV2 instance. This can be used to 'drop down' to the core SDK */
-  smartAccountClient: BiconomySmartAccountV2 | null
+  smartAccountClient: BiconomySmartAccountV2 | null;
   /** The Tanstack Query client instance */
-  queryClient: QueryClient | undefined
+  queryClient: QueryClient | undefined;
   /** The address of the smart account for the user */
-  smartAccountAddress: Hex
+  smartAccountAddress: Hex;
   /** The URL of the Biconomy bundler. This can be retrieved from the Biconomy dashboard: https://dashboard.biconomy.io */
-  bundlerUrl: string
+  bundlerUrl: string;
   /** The paymaster API key. This can be retrieved from the Biconomy dashboard: https://dashboard.biconomy.io */
-  biconomyPaymasterApiKey: string
-}
+  biconomyPaymasterApiKey: string;
+  /** A boolean indicating whether the smart account is loading */
+  smartAccountQuery: Omit<UseQueryResult, "data"> | null;
+};
 
 /** @ignore */
 export const BiconomyContext = createContext<BiconomyContextPayload>({
@@ -45,8 +48,9 @@ export const BiconomyContext = createContext<BiconomyContextPayload>({
   queryClient: undefined,
   smartAccountAddress: "0x",
   bundlerUrl: "",
-  biconomyPaymasterApiKey: ""
-})
+  biconomyPaymasterApiKey: "",
+  smartAccountQuery: null,
+});
 
 /**
 
@@ -89,30 +93,31 @@ ReactDOM.createRoot(document?.getElementById("root")!).render(
 );
 */
 export const BiconomyProvider = (props: BiconomyProviderProps) => {
-  const { children, config, queryClient } = props
-  const { bundlerUrl, biconomyPaymasterApiKey } = config
-  const [smartAccountClient, setSmartAccountClient] =
-    useState<BiconomySmartAccountV2 | null>(null)
-  const [smartAccountAddress, setSmartAccountAddress] = useState<Hex>("0x")
-  const { data: signer } = useWalletClient()
-  const switchAccount =
-    smartAccountClient?.biconomySmartAccountConfig?.signer?.inner?.account
-      ?.address !== signer?.account?.address
+  const { children, config, queryClient } = props;
+  const { bundlerUrl, biconomyPaymasterApiKey } = config;
+  const { data: signer } = useWalletClient();
 
-  useEffect(() => {
-    const createSmartAccount = async () => {
-      if ((!smartAccountClient || switchAccount) && !!signer) {
+  const { data, ...smartAccountQuery } = useQuery({
+    queryKey: ["smart-account-address", signer?.account.address],
+    queryFn: async () => {
+      if (signer) {
         const smartAccount = await createSmartAccountClient({
           signer: signer as SupportedSigner,
-          bundlerUrl, biconomyPaymasterApiKey
-        })
-        setSmartAccountClient(smartAccount)
-        setSmartAccountAddress(await smartAccount.getAccountAddress())
+          bundlerUrl,
+          biconomyPaymasterApiKey,
+        });
+        const smartAccountAddress = await smartAccount.getAccountAddress();
+        return {
+          smartAccount,
+          smartAccountAddress,
+        };
       }
-    }
+      return {};
+    },
+  });
 
-    createSmartAccount()
-  }, [signer, bundlerUrl, biconomyPaymasterApiKey, smartAccountClient, switchAccount])
+  const smartAccountClient = data?.smartAccount ?? null;
+  const smartAccountAddress = data?.smartAccountAddress ?? "0x";
 
   const value = useMemo(
     () => ({
@@ -120,20 +125,22 @@ export const BiconomyProvider = (props: BiconomyProviderProps) => {
       queryClient,
       smartAccountAddress,
       bundlerUrl,
-      biconomyPaymasterApiKey
+      biconomyPaymasterApiKey,
+      smartAccountQuery,
     }),
     [
       smartAccountClient,
       queryClient,
       smartAccountAddress,
       bundlerUrl,
-      biconomyPaymasterApiKey
+      biconomyPaymasterApiKey,
+      smartAccountQuery,
     ]
-  )
+  );
 
   return (
     <BiconomyContext.Provider value={value}>
       {children}
     </BiconomyContext.Provider>
-  )
-}
+  );
+};